+++ library/src/main/java/com/prolificinteractive/materialcalendarview/CalendarMode.java
@@ -1,0 +1,13 @@
+package com.prolificinteractive.materialcalendarview;
+
+public enum CalendarMode {
+
+    MONTHS(6),
+    WEEKS(1);
+
+    final int visibleWeeksCount;
+
+    CalendarMode(int visibleWeeksCount) {
+        this.visibleWeeksCount = visibleWeeksCount;
+    }
+}