+++ library/src/main/java/com/prolificinteractive/materialcalendarview/CalendarPager.java
@@ -7,19 +7,12 @@ import android.view.MotionEvent;
 /**
  * Custom ViewPager that allows swiping to be disabled.
  */
-class MonthPager extends BetterViewPager {
+class CalendarPager extends BetterViewPager {
 
     private boolean pagingEnabled = true;
 
-    public MonthPager(Context context) {
+    public CalendarPager(Context context) {
         super(context);
-    }
-
-    @Override
-    public void scrollTo(int x, int y) {
-        if (pagingEnabled) {
-            super.scrollTo(x, y);
-        }
     }
 
     /**
@@ -44,7 +37,7 @@ class MonthPager extends BetterViewPager {
     }
 
     @Override
-    public void setPageTransformer(boolean reverseDrawingOrder, PageTransformer transformer) {
-        super.setPageTransformer(reverseDrawingOrder, transformer);
+    public boolean onTouchEvent(MotionEvent ev) {
+        return pagingEnabled && super.onTouchEvent(ev);
     }
 }