+++ library/src/main/java/com/prolificinteractive/materialcalendarview/MaterialCalendarView.java
@@ -17,6 +17,7 @@ import android.util.AttributeSet;
 import android.util.SparseArray;
 import android.util.TypedValue;
 import android.view.Gravity;
+import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.accessibility.AccessibilityEvent;
@@ -146,6 +147,8 @@ public class MaterialCalendarView extends ViewGroup {
      * Default tile size in DIPs. This is used in cases where there is no tile size specificed and the view is set to {@linkplain ViewGroup.LayoutParams#WRAP_CONTENT WRAP_CONTENT}
      */
     public static final int DEFAULT_TILE_SIZE_DP = 44;
+    private static final int DEFAULT_DAYS_IN_WEEK = 7;
+    private static final int DAY_NAMES_ROW = 1;
 
     private static final TitleFormatter DEFAULT_TITLE_FORMATTER = new DateFormatTitleFormatter();
     private final TitleChanger titleChanger;
@@ -153,10 +156,11 @@ public class MaterialCalendarView extends ViewGroup {
     private final TextView title;
     private final DirectionButton buttonPast;
     private final DirectionButton buttonFuture;
-    private final MonthPager pager;
-    private final MonthPagerAdapter adapter;
+    private final CalendarPager pager;
+    private CalendarPagerAdapter<?> adapter;
     private CalendarDay currentMonth;
     private LinearLayout topbar;
+    private CalendarMode calendarMode = CalendarMode.MONTHS;
     /**
      * Used for the dynamic calendar height.
      */
@@ -228,7 +232,7 @@ public class MaterialCalendarView extends ViewGroup {
         buttonPast = new DirectionButton(getContext());
         title = new TextView(getContext());
         buttonFuture = new DirectionButton(getContext());
-        pager = new MonthPager(getContext());
+        pager = new CalendarPager(getContext());
 
         setupChildren();
 
@@ -338,7 +342,7 @@ public class MaterialCalendarView extends ViewGroup {
             monthView.setDateTextAppearance(adapter.getDateTextAppearance());
             monthView.setWeekDayTextAppearance(adapter.getWeekDayTextAppearance());
             monthView.setShowOtherDates(getShowOtherDates());
-            addView(monthView, new LayoutParams(MonthView.DEFAULT_MONTH_TILE_HEIGHT));
+            addView(monthView, new LayoutParams(calendarMode.visibleWeeksCount + DAY_NAMES_ROW));
         }
     }
 
@@ -356,7 +360,7 @@ public class MaterialCalendarView extends ViewGroup {
 
         title.setGravity(Gravity.CENTER);
         topbar.addView(title, new LinearLayout.LayoutParams(
-                0, LayoutParams.MATCH_PARENT, MonthView.DEFAULT_DAYS_IN_WEEK - 2
+                0, LayoutParams.MATCH_PARENT, DEFAULT_DAYS_IN_WEEK - 2
         ));
 
         buttonFuture.setScaleType(ImageView.ScaleType.CENTER_INSIDE);
@@ -365,7 +369,7 @@ public class MaterialCalendarView extends ViewGroup {
 
         pager.setId(R.id.mcv_pager);
         pager.setOffscreenPageLimit(1);
-        addView(pager, new LayoutParams(MonthView.DEFAULT_MONTH_TILE_HEIGHT));
+        addView(pager, new LayoutParams(calendarMode.visibleWeeksCount + DAY_NAMES_ROW));
     }
 
     private void updateUi() {
@@ -415,6 +419,39 @@ public class MaterialCalendarView extends ViewGroup {
         }
 
         adapter.setSelectionEnabled(selectionMode != SELECTION_MODE_NONE);
+    }
+
+    /**
+     * Set calendar display mode. The default mode is Months.
+     * When switching between modes will select todays date, or the selected date,
+     * if selection mode is single.
+     *
+     * @param mode - calendar mode
+     */
+    public void setCalendarDisplayMode(CalendarMode mode) {
+        if (calendarMode.equals(mode)) {
+            return;
+        }
+
+        CalendarPagerAdapter<?> newAdapter;
+        switch (mode) {
+            case MONTHS:
+                newAdapter = new MonthPagerAdapter(this);
+                break;
+            case WEEKS:
+                newAdapter = new WeekPagerAdapter(this);
+                break;
+            default:
+                throw new IllegalArgumentException("Provided display mode which is not yet implemented");
+        }
+        adapter = adapter.migrateStateAndReturn(newAdapter);
+        pager.setAdapter(adapter);
+        calendarMode = mode;
+        setCurrentDate(selectionMode == SELECTION_MODE_SINGLE
+                ? adapter.getSelectedDates().get(0)
+                : CalendarDay.today());
+        invalidateDecorators();
+        updateUi();
     }
 
     /**
@@ -470,7 +507,18 @@ public class MaterialCalendarView extends ViewGroup {
      * @return true if there is a future month that can be shown
      */
     private boolean canGoForward() {
-        return pager.isPagingEnabled() && pager.getCurrentItem() < (adapter.getCount() - 1);
+        return pager.getCurrentItem() < (adapter.getCount() - 1);
+    }
+
+    /**
+     * Pass all touch events to the pager so scrolling works on the edges of the calendar view.
+     * 
+     * @param event
+     * @return
+     */
+    @Override
+    public boolean onTouchEvent(MotionEvent event) {
+        return pager.dispatchTouchEvent(event);
     }
 
     /**
@@ -479,7 +527,7 @@ public class MaterialCalendarView extends ViewGroup {
      * @return true if there is a previous month that can be shown
      */
     private boolean canGoBack() {
-        return pager.isPagingEnabled() && pager.getCurrentItem() > 0;
+        return pager.getCurrentItem() > 0;
     }
 
     /**
@@ -1022,7 +1070,7 @@ public class MaterialCalendarView extends ViewGroup {
 
     /**
      * Sets the first day of the week.
-
+     * <p/>
      * Uses the java.util.Calendar day constants.
      *
      * @param day The first day of the week as a java.util.Calendar day constant.
@@ -1042,7 +1090,7 @@ public class MaterialCalendarView extends ViewGroup {
     /**
      * By default, the calendar will take up all the space needed to show any month (6 rows).
      * By enabling dynamic height, the view will change height dependant on the visible month.
-     *
+     * <p/>
      * This means months that only need 5 or 4 rows to show the entire month will only take up
      * that many rows, and will grow and shrink as necessary.
      *
@@ -1255,24 +1303,12 @@ public class MaterialCalendarView extends ViewGroup {
         final int desiredWidth = specWidthSize - getPaddingLeft() - getPaddingRight();
         final int desiredHeight = specHeightSize - getPaddingTop() - getPaddingBottom();
 
-        int weekCount = MonthView.DEFAULT_MONTH_TILE_HEIGHT;
-
-        /*
-         * The default height of the calendar component is MonthView.DEFAULT_MONTH_TILE_HEIGHT rows,
-         * but we need to hide 1 if the actual date is 5 week long
-         */
-        if (mDynamicHeightEnabled && adapter != null && pager != null) {
-            Calendar cal = (Calendar) adapter.getItem(pager.getCurrentItem()).getCalendar().clone();
-            cal.set(Calendar.DAY_OF_MONTH, cal.getActualMaximum(Calendar.DAY_OF_MONTH));
-            //noinspection ResourceType
-            cal.setFirstDayOfWeek(getFirstDayOfWeek());
-            weekCount = cal.get(Calendar.WEEK_OF_MONTH) + 1; // + 1 because of the week days
-        }
+        final int weekCount = getWeekCountBasedOnMode();
 
         final int viewTileHeight = getTopbarVisible() ? (weekCount + 1) : weekCount;
 
         //Calculate independent tile sizes for later
-        int desiredTileWidth = desiredWidth / MonthView.DEFAULT_DAYS_IN_WEEK;
+        int desiredTileWidth = desiredWidth / DEFAULT_DAYS_IN_WEEK;
         int desiredTileHeight = desiredHeight / viewTileHeight;
 
         int measureTileSize = -1;
@@ -1299,7 +1335,7 @@ public class MaterialCalendarView extends ViewGroup {
         }
 
         //Calculate our size based off our measured tile size
-        int measuredWidth = measureTileSize * MonthView.DEFAULT_DAYS_IN_WEEK;
+        int measuredWidth = measureTileSize * DEFAULT_DAYS_IN_WEEK;
         int measuredHeight = measureTileSize * viewTileHeight;
 
         //Put padding back in from when we took it away
@@ -1321,7 +1357,7 @@ public class MaterialCalendarView extends ViewGroup {
             LayoutParams p = (LayoutParams) child.getLayoutParams();
 
             int childWidthMeasureSpec = MeasureSpec.makeMeasureSpec(
-                    MonthView.DEFAULT_DAYS_IN_WEEK * measureTileSize,
+                    DEFAULT_DAYS_IN_WEEK * measureTileSize,
                     MeasureSpec.EXACTLY
             );
 
@@ -1332,6 +1368,19 @@ public class MaterialCalendarView extends ViewGroup {
 
             child.measure(childWidthMeasureSpec, childHeightMeasureSpec);
         }
+    }
+
+    private int getWeekCountBasedOnMode() {
+        int weekCount = calendarMode.visibleWeeksCount;
+        boolean isInMonthsMode = calendarMode.equals(CalendarMode.MONTHS);
+        if (isInMonthsMode && mDynamicHeightEnabled && adapter != null && pager != null) {
+            Calendar cal = (Calendar) adapter.getItem(pager.getCurrentItem()).getCalendar().clone();
+            cal.set(Calendar.DAY_OF_MONTH, cal.getActualMaximum(Calendar.DAY_OF_MONTH));
+            //noinspection ResourceType
+            cal.setFirstDayOfWeek(getFirstDayOfWeek());
+            weekCount = cal.get(Calendar.WEEK_OF_MONTH);
+        }
+        return weekCount + DAY_NAMES_ROW;
     }
 
     /**